EES-4716 Move Swagger config into Swagger namespace
