+++ src/GovUk.Education.ExploreEducationStatistics.Public.Data.Api/Swagger/SwaggerConfig.cs
@@ -1,0 +1,50 @@
+using Asp.Versioning.ApiExplorer;
+using Microsoft.Extensions.Options;
+using Microsoft.OpenApi.Models;
+using Swashbuckle.AspNetCore.SwaggerGen;
+
+namespace GovUk.Education.ExploreEducationStatistics.Public.Data.Api.Swagger;
+
+public class SwaggerConfig(IApiVersionDescriptionProvider provider) : IConfigureOptions<SwaggerGenOptions>
+{
+    public void Configure(SwaggerGenOptions options)
+    {
+        options.OperationFilter<DefaultValuesOperationFilter>();
+        options.SchemaFilter<JsonConverterSchemaFilter>();
+
+        var fileName = typeof(Program).Assembly.GetName().Name + ".xml";
+        var filePath = Path.Combine(AppContext.BaseDirectory, fileName);
+
+        options.IncludeXmlComments(filePath);
+        options.CustomOperationIds(apiDesc =>
+            {
+                var actionDescriptor = apiDesc.ActionDescriptor;
+
+                return actionDescriptor.AttributeRouteInfo?.Name
+                       ?? actionDescriptor.EndpointMetadata
+                               .OfType<IEndpointNameMetadata>()
+                               .LastOrDefault()
+                               ?.EndpointName
+                       ?? (apiDesc.TryGetMethodInfo(out var methodInfo) ? methodInfo.Name : null);
+            }
+        );
+
+        foreach (var description in provider.ApiVersionDescriptions)
+        {
+            options.SwaggerDoc(
+                description.GroupName,
+                new OpenApiInfo
+                {
+                    Title = "Explore education statistics - Public Data API",
+                    Version = description.ApiVersion.ToString(),
+                    Contact = new OpenApiContact
+                    {
+                        Name = "Explore education statistics",
+                        Email = "explore.statistics@education.gov.uk",
+                        Url = new Uri("https://explore-education-statistics.service.gov.uk")
+                    },
+                }
+            );
+        }
+    }
+}
